requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
braceexpand==0.1.7
Brotli @ file:///croot/brotli-split_1714483155106/work
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
coverage==5.5
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.9
distlib==0.3.9
docker-pycreds==0.4.0
dulwich==0.21.7
fairseq==0.12.2
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.12.0
ftfy==6.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
grpcio==1.62.1
h5py==3.12.1
huggingface-hub==0.27.1
hydra-core==1.0.7
idna==3.6
importlib_metadata==8.5.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.3.2
keyring==24.3.1
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.0.6
packaging==24.0
pandas==1.5.3
pathtools==0.1.2
pexpect==4.9.0
pidfile==0.1.1
pillow==10.2.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
-e git+ssh://git@github.com/keio-smilab24/Polos.git@12613990ae7cf336a39ee98fae034ba6bd2a6aaa#egg=polos
portalocker==2.8.2
protobuf==3.19.6
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pyparsing==3.1.2
pyproject_hooks==1.2.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytorch-lightning==1.2.10
pytorch-nlp==0.5.0
pytz==2024.1
PyYAML==5.3.1
RapidFuzz==3.11.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
sacrebleu==2.4.1
safetensors==0.5.2
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.3
sentry-sdk==1.42.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sphinx-markdown-tables==0.0.15
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.3.0
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.13.2
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchmetrics==0.2.0
torchvision==0.16.0+cu118
tqdm==4.66.2
transformers==4.46.1
triton==2.1.0
trove-classifiers==2024.10.21.16
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.28.0
wandb==0.15.12
wcwidth==0.2.13
webdataset==0.2.5
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.21.0
